+++ extensions/spotify/src/main/java/app/revanced/extension/spotify/misc/fix/SpoofClientPatch.java
@@ -11,20 +11,19 @@ public class SpoofClientPatch {
     /**
      * Injection point.
      * <br>
-     * Start login server.
+     * Launch login server.
      */
-    public static void listen(int port) {
+    public static void launchListener(int port) {
         if (listener != null) {
             Logger.printInfo(() -> "Listener already running on port " + port);
             return;
         }
 
         try {
+            Logger.printInfo(() -> "Launching listener on port " + port);
             listener = new LoginRequestListener(port);
-            listener.start();
-            Logger.printInfo(() -> "Listener running on port " + port);
         } catch (Exception ex) {
-            Logger.printException(() -> "listen failure", ex);
+            Logger.printException(() -> "launchListener failure", ex);
         }
     }
 
@@ -33,11 +32,11 @@ public class SpoofClientPatch {
      * <br>
      * Launch login web view.
      */
-    public static void login(LayoutInflater inflater) {
+    public static void launchLogin(LayoutInflater inflater) {
         try {
-            WebApp.login(inflater.getContext());
+            WebApp.launchLogin(inflater.getContext());
         } catch (Exception ex) {
-            Logger.printException(() -> "login failure", ex);
+            Logger.printException(() -> "launchLogin failure", ex);
         }
     }